+++ CHANGELOG.md
@@ -1,7 +1,16 @@
-## v0.11.0
+## v0.12.0
 
 - Add Cloudwatch dashboards configurations
+
+## v0.11.1
+
+- Removed TLSStore resource from ref.arch
+- Added seperate chain_mode boolean to deploy Traefik and Nginx in chain mode
+
+## v0.11.0
+
 - Added support for Traefik ingress controller in chaining mode
+- Add variable to make vpc-cni configurable
 
 ## v0.10.0